+++ src/Filters/Price/index.jsx
@@ -1,89 +1,49 @@
 import React from 'react';
-import Rheostat from 'rheostat';
 import styled from 'styled-components';
-import { letterFrequency } from '@vx/mock-data';
-import { Group } from '@vx/group';
-import { Bar } from '@vx/shape';
-import { scaleLinear, scaleBand } from '@vx/scale';
+import Modal from '../Modal';
+import InfoPanel from '../InfoPanel';
+import PriceSelector from './PureFilter';
 
-const Handle = styled.div`
-  background: white none;
-  border-radius: 100%;
-  border: 1px solid #008489;
-  cursor: pointer;
+const Wrapper = styled.div`
+  padding-left: 1rem;
+  padding-right: 1rem;
 `;
 
-const ProgressBar = styled.div`
-  background: #008489;
-  height: 6px;
-`;
+export default class PriceController extends React.Component {
+  state = {
+    min: 0,
+    max: 100,
+  };
 
-const Histogram = ({ className }) => {
-  const barsData = letterFrequency;
+  onChange = ({ values }) => {
+    const [min, max] = values;
+    this.setState({ min, max });
+  };
 
-  const width = 300;
-  const height = 100;
+  reset = () => {
+    this.setState({ ...this.props.getSavedState() });
+  };
 
-  const xMax = width;
-  const yMax = height;
+  save = () => {
+    this.props.onSave(this.state);
+    this.props.onClose();
+  };
 
-  const x = d => d.letter;
-  const y = d => +d.frequency * 100;
-
-  const xScale = scaleBand({
-    rangeRound: [0, xMax],
-    domain: barsData.map(x),
-  });
-  const yScale = scaleLinear({
-    rangeRound: [yMax, 0],
-    domain: [0, Math.max(...barsData.map(y))],
-  });
-
-  const compose = (scale, accessor) => data => scale(accessor(data));
-  const xPoint = compose(xScale, x);
-  const yPoint = compose(yScale, y);
-
-  return (
-    <svg width={width} height={height} className={className}>
-      {barsData.map((bar) => {
-        const barHeight = yMax - yPoint(bar);
-        return (
-          <Group key={`bar-${x(bar)}`}>
-            <Bar
-              x={xPoint(bar)}
-              y={yMax - barHeight}
-              height={barHeight}
-              width={xScale.bandwidth()}
-              fill="#D8D8D8"
-            />
-          </Group>
-        );
-      })}
-    </svg>
-  );
-};
-
-const HistogramAligned = styled(Histogram)`
-  margin-bottom: -6px;
-`;
-
-const PriceRange = styled.p`
-  font-weight: 100;
-  margin-top: 1rem;
-  margin-bottom: 0.5rem;
-`;
-
-const Description = styled.p`
-  font-weight: 100;
-  font-size: 0.85rem;
-  margin-bottom: 1rem;
-`;
-
-export default ({ onChange, min, max }) => (
-  <React.Fragment>
-    <PriceRange>$10 – $1000+</PriceRange>
-    <Description>The average nightly price is $75.</Description>
-    <HistogramAligned />
-    <Rheostat handle={Handle} progressBar={ProgressBar} onChange={onChange} values={[min, max]} />
-  </React.Fragment>
-);
+  render() {
+    return (
+      <Modal
+        {...this.props}
+        title="Price"
+        buttonTitle="Price"
+        onReset={this.reset}
+        onSave={this.save}
+        className="hidden-xs hidden-sm hidden-md"
+      >
+        <Wrapper>
+          <PriceSelector onChange={this.onChange} min={this.state.min} max={this.state.max} />
+        </Wrapper>
+        <InfoPanel onCancel={this.props.onCancel} onApply={this.save} />
+      </Modal>
+    );
+  }
+}
